Store coordinates as percentage

+++ db/migrate/20170601134521_create_tags.rb
@@ -1,8 +1,8 @@
 class CreateTags < ActiveRecord::Migration[5.0]
   def change
     create_table :tags do |t|
-      t.integer :x
-      t.integer :y
+      t.float :x
+      t.float :y
       t.references :character, foreign_key: true
 
       t.timestamps
